Move container info classes to separate file
