+++ nano/lib/CMakeLists.txt
@@ -36,6 +36,8 @@ add_library(
   config.hpp
   config.cpp
   configbase.hpp
+  container_info.hpp
+  container_info.cpp
   diagnosticsconfig.hpp
   diagnosticsconfig.cpp
   env.hpp
