+++ nano/lib/container_info.hpp
@@ -1,0 +1,50 @@
+#pragma once
+
+#include <string>
+#include <vector>
+
+namespace nano
+{
+/* These containers are used to collect information about sequence containers.
+ * It makes use of the composite design pattern to collect information
+ * from sequence containers and sequence containers inside member variables.
+ */
+struct container_info
+{
+	std::string name;
+	size_t count;
+	size_t sizeof_element;
+};
+
+class container_info_component
+{
+public:
+	virtual ~container_info_component () = default;
+	virtual bool is_composite () const = 0;
+};
+
+class container_info_composite : public container_info_component
+{
+public:
+	container_info_composite (std::string const & name);
+	bool is_composite () const override;
+	void add_component (std::unique_ptr<container_info_component> child);
+	std::vector<std::unique_ptr<container_info_component>> const & get_children () const;
+	std::string const & get_name () const;
+
+private:
+	std::string name;
+	std::vector<std::unique_ptr<container_info_component>> children;
+};
+
+class container_info_leaf : public container_info_component
+{
+public:
+	container_info_leaf (container_info const & info);
+	bool is_composite () const override;
+	container_info const & get_info () const;
+
+private:
+	container_info info;
+};
+}
